topic/#19

+++ src/main/java/com/example/management/model/Employee.java
@@ -38,7 +38,7 @@ public class Employee implements UserDetails {
 	
 	private LocalDateTime updatedAt; // 更新日時
 	
-	private Integer invalid; // 削除フラグ
+	private Short invalid; // 削除フラグ
 
 	/**
 	 * 社員の所持する権限を返すメソッド
