+++ src/main/java/com/example/management/model/Item.java
@@ -26,5 +26,5 @@ public class Item {
 	
 	private LocalDateTime updatedAt; // 更新日時
 	
-	private Integer invalid; // 削除フラグ
+	private Short invalid; // 削除フラグ
 }
